+++ packages/@cldcvr/flow-form-builder/src/components/f-form-builder/fields/file.ts
@@ -11,6 +11,8 @@ export default function (
 	const field = _field as FormBuilderFileField;
 	return html`
 		<f-file-upload
+			id=${ifDefined(field.id)}
+			class=${ifDefined(field.className)}
 			name=${name}
 			data-qa-element-id=${field.qaId || field.id}
 			${ref(fieldRef)}